// src/app/services/iex/interfaces/quote.ts
export interface Quote {
  symbol: string;
  companyName: string;
  primaryExchange: string;
  sector: string;
  calculationPrice: string;
  open: Number;
  openTime: Number;
  close: Number;
  closeTime: Number;
  high: Number;
  low: Number;
  latestPrice: Number;
  latestSource: string;
  latestTime: string;
  latestUpdate: Number;
  latestVolume: Number;
  iexRealtimePrice: Number;
  iexRealtimeSize: Number;
  iexLastUpdated: Number;
  delayedPrice: Number;
  delayedPriceTime: Number;
  extendedPrice: Number;
  extendedChange: Number;
  extendedChangePercent: Number;
  extendedPriceTime: Number;
  previousClose: Number;
  change: Number;
  changePercent: Number;
  iexMarketPercent: Number;
  iexVolume: Number;
  avgTotalVolume: Number;
  iexBidPrice: Number;
  iexBidSize: Number;
  iexAskPrice: Number;
  iexAskSize: Number;
  marketCap: Number;
  peRatio: Number;
  week52High: Number;
  week52Low: Number;
  ytdChange: Number;
}